Fix the "No device found." error when switching channels.
libvlc forks another thread. This is a problem when trying to close file
descriptors, because these can only be closed if all child-threads are
finished.
The correct fix for this is to simply set O_CLOEXEC when opening the
file descriptor of the frontend.
This lets the dvbdevice-thread close the handle again.

REVIEW: 106259