Get rid of dvbdevice_linux_interface ugly interface
dvbdevice_linux_interface.h and dvbdevice_linux_interface.cpp
came from the KDE4 solid, but the .cpp doesn't build. It is ugly,
as it were using some Solid internals.

Get rid of it by using the generic Solid notifications and adding
a logic there to handle the DVB specifics.

The logic assumes that the DVB devs are under /dev/dvb/adapter?.
This might not be the case, if someone hacks udev, but I've no
idea if there's a better way of doing that.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>