dvbconfigdialog: increase the size of the TV configuration
When more than one device is present, the size of the config
box is too small to show even the first device tab. Ok, there
will be some buttons to allow seeing the other tabs, but
this is not too practical.

So, increase it in a way that fits up to 4 devices. This
should cover most of the typical cases.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>