dvbliveview: Update the URL depending of the timeshift status
When the DVB is playing live view, the play backend should use
dvbpipe.m2t URL, but, if the stream gets paused, a timeshift
file will be recorded. So, when unpaused, the backend should use
the timeshift file instead.

So, update the URL depending if it is either in live view or
in timeshift mode.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>