vlcmediawidget: Update status asynchronously
The abstractmediawidget current model assumes that the status
changes are synchronous. This is not true with VLC, as it
takes some time for VLC to buffer the data and start playing
a new URL.

So, change the model to identify if the status really has
changed.

With this change, plus the last one, pause/play a DVB stream
now works properly.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>