dvbrecording: don't use epgModel when app is closing
When the Kaffeine application is being closed, it will
destroy all DvbRecordingFile instances, with, in turn,
calls its stop() method.

The logic at DvbRecordingFile::stop() will call
manager->getRecordingModel()->findNewRecordings() in
order to see if are there any other filters.

However, the epgModel is destroyed before DvbRecording.

So, the epgModel var will be pointing to an already freed
data, causing the application to crash.

Solve that by setting epgModel to NULL at the
DvbEpgFilter::~DvbEpgFilter() destructor and checking
if epgModel is NULL when trying to find for new records
at the EPG guide.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>