dvb: flush channel lists when the channel scan box is closed
Currently, the channel scan list is only saved when Kaffeine is
closed without errors. This is bad, as scanning for channels
take a while. So, let's flush the channel list when the "Ok"
button is clicked at the Channel Scan dialog. This way, if
something gets bad, the channel lists will be there.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>