vlcmediawidget: Improve the mouse unhide mechanism
There are still some troubles with the workaround for the
libVLC hide bug. So, change the logic to solve those issues.

Now, the hideMouse() function will not hide the mouse if
its position is outside the visible area of the play widget.

We're also getting rid of the setOverride stuff, as this
is not safe enough.

BUG: 364016

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>