dvbconfigdialog: scantable.dvb.qz is now hosted at linuxtv.org
As the dtv-scan-tables ML is hosted at linuxtv.org, let's host
there the scanfile.dvb.qz too. This way, the same script that
generate daily tarballs when patches are applied to
dtv-scan-tables will create a new scantable automatically.

That warrants that Kaffeine will aways use an updated version
of the channel lists.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>