dvbdevice_linux: reorder C includes and use extern "C"
We don't want those C includes to use the more restricted
C++ assumptions, as adding new items on enums, for example,
is something that we do from time to time at the Kernel
headers. A new value on an enum, in C++, may cause it to
crash.

While here, reorder the headers.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>