Better handle *.cmake at .gitignore
instead of including *.cmake at the .gitignore root and
explicitly excluding src/config-kaffeine.h.cmake at the
src/.gitignore, let's reverse the logic: exclude only the
*.cmake files that are auto-generated.

The advantage is that, if we add other files to be generated
via configure_file(), we won't need to exclude it. It also
helps to identify if something new with the *.cmake extension
is created without being expected.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>