vlcmediawidget: fix subtitles handling
The mediawidget assumes that subtitles are numbered starting from zero
and at the same order as they're stored at the QStringList. VLC,
however, has its own IDs that usually don't start with zero.

So, add a QMap to allow converting to/from those two different
convertions.

With that, subtitles are now properly shown and can be properly
selected.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>