vlcmediawidget: better handle subtitle names
Kaffeine currently expects subtypes to be in the form of:
	Language - [Lingua]
Like:
	English - [Inglês]
	Italiano - [Italiano]

And gets just the translated version of the language. However,
external file tracks are like:
	track %d + some comments

Where track is actually translated to the current language,
and the comment itself can be enclosed by []. So, let's
remove the cleanup heuristics, and just present the string
as-is to the user. If empty, we can then use:
	Subtitle %d

And let it to be translated.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>