mediawidget: Go to the next URL on a playlist at the end of a play
The logic of automatically go to the next item on a playlist is
currently broken.

When VLC detects the end of a media play, it emits a signal
(libvlc_MediaPlayerEndReached), with is returned to the media widget
as a playbackFinished signal. Such signal should cause a call to
PlaylistTab::playNextTrack() for its logic to play the next file.

BUG: 364224

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>