icons: fix tools/update_icons.sh to generate a proper CMakeLists.txt
Changeset 691997e3a6dd ("icons: use scalable icons also for encrypted radio/TV")
reimplemented the encrypted icons, but it forgot to change the
CMakeLists.txt.

This cause the icons to be installed on a wrong place. The
breakage was already fixed by changeset aff5fab71f9d
("Fix icon installation. All icons are scalable"), but we need
to also update the script, as otherwise if we run it again,
it will cause breakages.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>