icons: Don't install the icons at the hicolor dir
All the icons used by caffeine are now part of the kaffeine.qrc
resource file. That actually means that Qt5 rcc will generate
a file (src/qrc_kaffeine.cpp) that will have all icons statically
defined there, and will be linked with the other object files
to produce the Kaffeine executable.

As such, we don't need anymore to copy those icons to the
<prefix>/share/icons/hicolor/ directory.

So, let's remove the CMakeLists.txt from the icons directory,
and let's not generate it anymore with tools/update_icons.sh
script. For now, I'll keep the logic that produces the
icons/CMakeLists.txt, as it might be useful on some future.

I'll eventually cleanup the script after some time.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>