mainwindow: allow the user to control debug verbosity
Right now, qInfo() has been used for all types of messages,
including debug ones. That's because, on Fedora 23, the
debug messages are disabled by design:
	https://bugzilla.redhat.com/show_bug.cgi?id=1227295

(and Fedora is installed on the machine I'm developing Kaffeine)

That's bad, since we want to allow the user to see some messages,
in order to help us to fix some hard to track bugs.

So, let's handle ourselves the debug verbosity by adding an
extra parameter. This way, we can start using the different
levels of verbosity, making Kaffeine quiet on normal usage,
but letting the user enable verbosity when reporting bugs.

Next patches will be changing the current verbosity and adding
some extra qDebug() on places where it is needed.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>