dvbscan: fix SDT entries addition
The current logic is broken because it assumes that all SDT entries
will have a 0x48 descriptor. As this is not always true, it ends by
creating invalid entries like:
	15-06-16 21:54:02.558 [Debug   ] New SDT entry: name , provider

It also assumes that one SDT entry has just one 0x48 descriptor.
This is usually the case, but I won't be surprised to find multiple
ones on some MPEG-TS, specially on DVB-C, where providers may
do whatever they want.

So, let's move the logic that adds the SDT entries to the place
where they're found.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>