dvbscan: fix handling of multiple sections count
The number of multiple sections is not global inside a
table. Instead, it is per-extension.

Fix the code that checks if handling a table is finished,
by making it to take into account the extension number.

With that, it is now possible to add support for other NIT
and other SDT tables.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>