dvbsi: fix handling of ISDB-Tb strings with LS0/LS1 codes
Some strings for ISDB-Tb use <LS1>string<LS0> format from
the latin extension table. As 0x0e (LS1) code is an invalid
code at the ETSI EN 300 468 Annex A, use it to identify
a code following the Latin charset.

We should also remove the final <LS0> char, as otherwise, it
will add an extra symbol at the end of the strings.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>