dvbepg: Add more content strings for ISDB-Tb and fix some strings
I was lacking the English version of ABNT NBR15603-2, with contains
the Annex C with the English strings for the content types.
Now that I got access to the content of such table, I can fix the
translations for the existing strings, and add the missing nibble2
table.

So, now, EPG parsing for ISDB-Tb should match the specs.
We still need to fix support for the version used in Asia,
as we need to detect the Country code either from the EPG
tables or from the system Contry definitions.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>