dvbscandialog: better callibrate the scale for the gauge
Instead of always use 0-100, with fits well for percentages,
use a ranges for signal and S/N, in a way that it will be
green when signal and S/N is good.

Please notice that it is possible to have values out of
such range. In that case, it would mean either 0% of the
gauge filled (if value is lower than the minimal) or 100%
of the gauge filled, if the value is higher than the max.

While here, remove the extra uneeded "{}" at the paintEvent
function.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>