dvbliveview: get rid of internal subtitles logic
libVLC has already a logic that handles subtitles. It
uses a different nomenclature than the one at this file,
with makes really hard to make it do the right thing, if
we keep doing the same thing on both places.

So, get rid of the logic here, and rely on libVLC.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>