Revert changeset f825ca8184ef1
Kaffeine 1.x used to have a hack to support a proprietary
userspace driver, with cheats the system for it to think that
there's a Kernel driver, emulating Kernel syscalls on userspace.

This hack never worked on Kaffeine 2.x, as shown on bug
365083.

As we're about to move device detection to libdvbv5, with
relies on other Kernel mechanisms that are not emulated by
such proprietary driver, let's remove the hack.

This change alone, removes 12% of useless code at
dvbdevice_linux.cpp, with is a good thing.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>