CMakeLists.txt: fix build breakage with ki18n 5.34.0
As reported on BZ#379834, Kaffeine doesn't build anymore with
ki18n. Add the suggested workaround, until we find some time to
make it compatible with the new cmake policies.

Please notice that here, with kf5-ki18n-5.34.0-2.fc25.x86_64 and
cmake-3.8.0-3.fc25.x86_64, I was unable to reproduce the bug.

However, as the workaround won't hurt, let's apply it.

BUG: 379834
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>