dvbliveview: better handle errors when writing to pipe
The current logic of writeToPipe() loops forever if an error
occurs. That causes CPU load to be 100%, as it will call write()
in non-block mode inside a loop until it succeeds. That's not
needed, as the write logic can already monitor the file descriptor
availability via QSocketNotifier.

So, modify the logic to rely on it.

Still, for some reason, it may retry too many times, causing
the buffer data to be too old to be meaningful. On such cases,
let's restart the pipe and hope for the best.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>