dvbliveview: simplify retry logic
We don't need an extra var to tell if the stream is paused or
not, as we can test if the timeshift file is opened. Also,
it shoudn't make any difference if we reset the pipe or not.
So, just print a warning if writing to the pipe is causing too
much retries.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>