mediawidget: don't use key sequences as shortcut
QKeySequence is used to define a sequence of keys that should
be pressed in order for an action to occur, and not to set
alternate keys for an action.

Due to that, the default key maps for several actions don't
work.

Also, on modern desktops, the Keyboard's media keys are already
mapped by the window manager, with may prevent Qt to use them.

So, instead, just use keystrokes that exist on all keyboards,
as, if someone wants something different, it can change the
shortcuts via Kaffeine's menu.

BUG: 383332

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>