dvbepg: don't merge messages from different languages
Some countries have multiple EPG entries for the same
program, with different languages. There's the case, for
example, of Finland, as reported on bug #385386.

We shouldn't be mixing different languages on the same
entry, as this is really messy.

So, add support to get the EPG languages.

For now, get just the first language, ignoring all
other ones.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>