dvb: add limited support to display EPG on multiple languages
Add the needed bits for storing multi language EPG at the EPG
database.

When displaying EPG info, use values for all languages.

This is not optimal, as users will likely want to see the
EPG on their own language instead of on some alien one, but
language selection can be added on a separate patch.

It should be noticed that, due to database changes, people
using the old database may notice duplicated values for a
while, after migrating to the new database version, as the
old database was not storing the EPG language.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>