dvbepg: ignore invalid language code
If the language code starts with 0x00 or 0x20, it means that
it is a blank/invalid code. So, ignore it, and handle as just
the "first" language.

While here, optimize the logic that handles the language
code, placing it on one ancillary function that it is
called where needed. That makes the logic simpler.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>