dvbepg: don't display "first:" if just one language is available
when code == "first", it means that the code information was not
filled on a given EPG entry, with should mean that there's just
one language for such entry.

So, there's no sense of displaying the "first:" on those.

The same applies to the EPG language select: it doesn't make
sense to present "first" as one possible languages.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>