mediawidget: switch DVD navigate to keyboard
With newer Qt classes and newer libVLC, it is not possible anymore
to navigate a DVD menu with the mouse.

The reason is that Qt libs intercept all Linux input events,
not letting libVLC to receive them.

The solution would be to let Kaffeine to pass mouse events directly
to libVLC. Unfortunately, such API doesn't exist. There was a set
of RFC patches adding support for it:

	http://vlc-devel.videolan.narkive.com/ah5I2Gpy/patch-libvlc-add-api-to-propagate-mouse-move-and-down-up-events

But this was never accepted upstream.

So, while this doesn't work, let's fall back to a different way:
DVD menu navigation should be done via keyboard interaction.

Basically, navigation is done via Up/Down/Left/Right keys. Once the
DVB menu item is selected, pressing <ENTER> will activate the item.

The libVLC provides support for a "Popup" navigate interaction, but
I was unsure how this is used. So, support for it was not implemented.

BUG: 373814

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>