mainwindow: ensure that the display mode is properly applied
When Kaffeine isn't in normal mode, activating a new tab won't
work, as the code with sets the active tab is not called. It
should be called, no matter what.

Also, emit a displayModeChanged() signal there, to be sure that
the menu hide/unhide will be properly set, and that the cursor
hiding time will restart.

BUG: 393533

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>