vlcmediawidget: Set vlcMedia to NULL after release
As we're now calling the code at makePlay() on two different
places, we should ensure that it will only work if vlcMedia
was created.

Without that, if someone tries to play an audio CD and it
fails for whatever reason (like if the system doesn't have
a CD device), it crashes.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>