vlcmediawidget: handle errors inside playDirection()
Move the code with fixes the trackNumber to a better place,
as, when using a new Audio CD track number fails, the logic
with handles it should be at playDirection(), with is the
one that actually increments it.

This is better than just reset trackNumber to 1, specially
if we end by adding support for replay() for Audio CD sources.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>