README.md: Split packaging info from prerequisites
Right now, the instructions are messy, as it mixes Debian/Ubuntu
packaging with build prerequisites.

So, move the instructions about how to build a .deb package
to another section.

We might need to change it in the future, if we decide to
also add instructions to build RPM files. I may eventually do
that some day.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>