vlcmediawidget: include vlc.h instead of defining classes directly
Clang 7 warns about some of the vlc.h specific class/structs, when built
with vlc 3. The reason is that some of those are not declared as
class, but, instead, as struct.

We might just fix one by one to match vlc 3 header, but this could
have changed in the past and may change in the future. So, just include
the header.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>