dvbchanneldialog: fix audio channel selection
While Kaffeine is properly detecting the audio channels inside
a MPEG-TS stream, the selection dialog is disabled.

The problem is that Kaffeine's logic is to allow selecting the
audio channel only if there are more than one audio PID, but
the check is at the wrong place!

So, move it to be at the right place. While here, do the check
using the checkbox variable, instead of relying on something
else.

BUG: 402735

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>