dvbtab: replay a channel if the PID settings gets updated
The DVB tab display allows to edit a channel that is being
played.

One of the possible changes there is to alter the audio
PIDs (and, maybe the video PID, if we end needing it).

Right now, such change doesn't affect the currently played
stream, with seems counter-intuitive, as all other control
changes affects runtime.

So, change the logic to detect such case and apply the new
changes, if a Digital TV PID was changed.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>