xmltv: add a parser for XMLTV EPG data files
This parser handles XMLTV format files as specified at:
	http://wiki.xmltv.org/index.php/Main_Page

The way the parser was written, once a file is added to
the parser, it keeps monitoring it for file changes.

When a file change is detected, it will reload the file and
parse it, adding entries to the EPG guide.

BUG: 405278

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>