xmltv: associate a xmltv channel to multiple DVB channels
It is not uncommon that a single TV channel to have multiple
video streams, with different resolutions (SD, HD, 4K, 8K).
On those cases, the same EPG data should be added to the
different video streams. So, be sure that it will try to
use the xmltv data for all possible display-names.

BUG: 405278

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>