xmltv: simplify XML parsers and fix it
There was a bug here that was causing problems with tags that
were both start and end elements, e. g. :
	<icon .... />

Simplify the logic used to parse the several XML fields,
in order to make the code simpler to read and check if the
end tag matches the start tag for each xml hierarchical
parser.

While here, add warnings when a new tag that can't be
parsed is found.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>