xmltv: improve category handle to match DTD spec
From DTD spec, the category can have a language.

This way, when supported, a multi-language EPG will
have translated category names.

Improve the parser in order to handle it.

What we detected, in practice, is that, at least on
some grabbers, all EPG entries are on one language,
while the category are in English. A dumb parser
would just create a separate entry for the category
using a different language.

Our parser, however is smart enough to identify if there's
just one language. In this case, even if the category
language is specified, it will simply ignore it.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>