dvbdevice: fix a race condition with frontend timeout
There's a race condition around frontendTimeout logic:
when the device is in idle state, a new tune may occur.

That will change the autoTransponder to point to the newly
channel in tune. So, we need to invalidate the autoTransponder
*before* changing the state to DeviceIdle.

Failing to do that makes Kaffeine to invalidade the next
channel, with causes warnings like those:

	30-04-19 12:06:53.952 [Warning ] kaffeine.dev: Invalid transmission type when get props from 497.14 MHz
	30-04-19 12:07:06.216 [Warning ] kaffeine.dev: Invalid transmission type when get props from 515.14 MHz
	30-04-19 12:07:12.570 [Warning ] kaffeine.dev: Invalid transmission type when get props from 527.14 MHz

With may end by making it not to add the proper tuning
parameters for good channels.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>