dvbtab: remove unused hide cursor logic
When the support for minimal mode and full screen mode to access the
lateral bar was added, it came with a auto-hide code. That was because
the code was copied from the main window.

Yet, such code is not needed and even produces a warning:

	01-05-19 00:08:08.138 [Warning ] QObject::connect: No such slot DvbTab::hideCursor() in /devel/v4l/kaffeine/src/dvb/dvbtab.cpp:229

So, get rid of it.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>