vlc: fix support for vlc 2.x
As reported by Wolfgang, the changes made on Kaffeine 2.0.16 to
solve issues with audio CDs broke Kaffeine when building it with
legacy vlc 2.2.x.

The fix here is simple: just place two libVlc 3.0 events at the
print logic inside the Vlc3 block is enough to make it build
again.

The libvlc_MediaMetaChanged is used to report when a new media is
playing.

With vlc 2.0, such event should be enabled only inside mediaPlay()
routine. Doing it early causes the play to not work.

So, re-add such logic, with got removed on changeset d03abc77ad40
("backend-vlc: simplify events handling logic").

BUG: 407153

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>