mediawidget:fix the mouse wheel behavior at the streaming window
The mouse vertical wheel is context sensitive, having different
functions depending on what part of Kaffeine's screen is used.

After changeset  95a76b59af9b ("mainwindow: use mouse wheel
to control volume"), the expected behavior is to control the
audio volume.

That works fine on live view. However, when this feature got
implemented, I didn't notice that, for DVD and other audio/video
plays, the wheel were used to control the position.

Due to that, currently, trying to change the wheel while playing
a video, CD or DVD will cause it to change both volume and position.

As it makes a lot more sense to adjust volume via the wheel, let's
drop support for changing the video position using it.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>