Add a Dockerfile to allow building Kaffeine inside a container
This is a basic Dockerfile with builds Kaffeine inside an
Ubuntu Disco container.

With the current settings, the container doesn't access the
DVB devices with are outside the container.

A command that would allow opening a file from your home
directory and to use a Digital TV device would be like:

    $ docker run -it --env="DISPLAY" --env="QT_X11_NO_MITSHM=1" --volume="/tmp/.X11-unix:/tmp/.X11-unix:rw" --volume="$HOME:$HOME:rw" $(for i in $(ls /dev/dvb/adapter*/*); do echo -n "--device=$i:$i:rwm "; done) kaffeine

Please notice, however, that, with the above there won't be
any persistent storage. every time the docker container
runs, it will forget any previously configured data, EPG
information, etc.

BUG: 388232

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>