Fix CI build on FreeBSD due to duplicate definitions
FreeBSD has linux/dvb/dmx.h (from the v4l port, which we require for
building kaffeine -- this is packaging intersecting with porting),
which has all the DVB definitions. Having an extra copy in include/dmx.h
is confusing, since there are different paths to get to the header,
and they use different header guards. This leads to duplicate
definitions and a build failure.

Packaging has "fixed" this with rm -rf include/ (in the kaffeine
sources) for ages.

Here, recognize that the system-provided dbx.h is the right one,
look for it with find_path() and use it if possible.