dvbsi: improve stream type detection logic
There are two ways of detecting the type of a stream:
use the stream type and using the registration descriptor.

The problem with stream type is that types like 0x06
can be used by different types of ES.

Some formats like SMPTE 302m can only be identified via
the registration descriptor.

Add support for it.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>