dvb: re-write bcdToTime function
The current bcdToTime logic has a few issues. The first one
is that the event time cannot be greater than 23:59:59, as
otherwise the EPG entry will become invalid.

Also, if either minutes or seconds are bigger than 59,
it will also produce an invalid entry. While this shouldn't
happen, in practice, the ET 300 468 doesn't forbid it.

Yet, as the values there are in BCD format, the maximum
possible value is actually 99:99:99.

So, better to be safe than sorry, and change the logic to
accept any valid BCD values.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>