dvbconfigdialog: increase max download size for scandata
By the time Kaffeine was written, a 64KB max limit were
enough, but nowadays, the scandata file has already 104KB.

Increase the limit up to 1MB, in order to fix download
issues with valid files.

While here, also improve the download messages to show
how much data was downloaded and to report the reason
why a download was aborted.

BUG: 436371

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>