main: restore backward compat with  KF < 5.62
Commit 4280d21a96c4 ("Port away from API deprecated in >KF-5.62")
broke compatibility with older KF versions.

That caused both the Travis CI job and Coverity at
https://builder.linuxtv.org/job/kaffeine-coverity/
to break.

While Travis CI was using an Ubuntu version that it entered
on extended LTS support, the builder uses the latest current
Debian LTS version. So, it seems worth enough to fix it,
instead of hacking the building environment.

So, restore the old behavior by  adding a version check.

Fixes: 4280d21a96c4 ("Port away from API deprecated in >KF-5.62")
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>